Initial Commit
This is the first commit of the code for the GLTF Unreal Importer.  It is based on a mixture of the USD and FBX file importers. There is still a lot of code commented out which was taken directly from the USD or FBX importers. This code will be changed or removed based on the ongoing work for the GLTF importer. The code is also not refactored in anyway for the GLTF loading. These files will be changing and could possibly be removed, renamed or completely re-written.