Loading in binary glb files with textures
Textures are being loaded in directly from the buffers. So at the moment there is wasted processing time loading in and converting to RGB using the  stb_image.h.

Currently stb_image.h is not required at all by this importer, and its just adding an additional overhead to the load times. Unless I find a way to convert from rgb to unreals internal file format. Or possibly write an image importer that takes raw "RGB" and "RGBA" files as input.