Glossiness is now added to roughness input. Specular has been disabled.
The alpha channel from specularGlossinessTexture map is inverted (OneMinus) and then multiplied against the glosinessFactor and input into the roughness channel of the material.

Specular (rgb component of specularGlossinessTexture map) has currently been disabled since it was being added to the specular input of a material but this caused the material to look too dark. So for now specular maps are not used.