Login Window closing was causing issues
Previous code still kept an invisible window around that was recieving clicks. This code fixes that issue, but I also had to remove the checks if the window was already open, since the flags to check if its closed is not working yet. So for now you can have multiple windows open.